Разобрала редьюсер на слайсы
