+++ src/components/main-map.tsx
@@ -2,9 +2,8 @@ import {useRef, useEffect} from 'react';
 import leaflet, {layerGroup, Marker } from 'leaflet';
 import 'leaflet/dist/leaflet.css';
 import { TypeOffer } from '../types/types-data';
-import { URL_MARKER_CURRENT, URL_MARKER_DEFAULT } from '../const';
+import { DEFAULT_CITY, URL_MARKER_CURRENT, URL_MARKER_DEFAULT } from '../const';
 import useMap from './use-map';
-import { DEFAULT_CITY } from '../store/reducer';
 
 
 type ListProps = {
