+++ src/store/index.ts
@@ -1,10 +1,10 @@
 import { configureStore } from '@reduxjs/toolkit';
-import { reducer } from './reducer';
 import { createAPI } from '../services/api';
+import { rootReducer } from './root-reducer';
 
 export const api = createAPI();
 export const store = configureStore({
-  reducer,
+  reducer:rootReducer,
   middleware:(getDefaultMiddleware)=>
     getDefaultMiddleware({
       thunk:{
